+++ src/appleseed/renderer/kernel/lighting/lighttypes.h
@@ -75,9 +75,9 @@ class EmittingTriangle
     size_t                      m_object_instance_index;
     size_t                      m_triangle_index;
     size_t                      m_light_tree_node_index;
-    foundation::Vector3d        m_v0, m_v1, m_v2;               // world space vertices of the triangle
-    foundation::Vector3d        m_n0, m_n1, m_n2;               // world space vertex normals
-    foundation::Vector3d        m_geometric_normal;             // world space geometric normal, unit-length
+    foundation::Vector3f        m_v0, m_v1, m_v2;               // world space vertices of the triangle
+    foundation::Vector3f        m_n0, m_n1, m_n2;               // world space vertex normals
+    foundation::Vector3f        m_geometric_normal;             // world space geometric normal, unit-length
     TriangleSupportPlaneType    m_triangle_support_plane;       // support plane of the triangle in assembly space
     float                       m_area;                         // world space triangle area
     float                       m_rcp_area;                     // world space triangle area reciprocal
