+++ src/appleseed/renderer/kernel/rendering/isamplerenderer.h
@@ -64,7 +64,7 @@ class ISampleRenderer
     virtual void render_sample(
         SamplingContext&                sampling_context,
         const PixelContext&             pixel_context,
-        const foundation::Vector2d&     image_point,
+        const foundation::Vector2f&     image_point,
         AOVAccumulatorContainer&        aov_accumulators,
         ShadingResult&                  shading_result) = 0;
